Adds worker model, spec, factory. Adds Seed file
+ Why is this change necessary?
It adds the model and validations for apollo workers. This is necessary
to generate the access_token for each worker.
+ How does this change address the issue?
It adds a model and instance method to generate and save the
access_token. It also adds a factory and model validations.

The seed file just calls the Series and Episode Builder service build
all methods. This was implemented to save dev time when rebuilding the
DB.

+ What side effects does this change have?
None.